+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Implement `xsd:NCName` data type codec
 - Implement `xsd:short` data type codec
 - Implement `xsd:byte` data type codec
+- Implement `xsd:base64Binary` data type codec